Add declaration of ORCHID color, but is it really this hex value?
So with recent addition of more tiers, up to 8, there was introduced a color ORCHID
which was not declared in surftimer already.
Next unused hex value out of those available is 0x1A, so I will take it as orchid.
Let us know if the color really is orchid ingame!

Without this declaration surftimer will fail to compile at all.