work around for style variables not being reset on mapchange
Styles variables don't get reset on mapchange, so if the next map has no style records, it will compare to prev map
Sometimes it shows rank 2/1, this happens because runtime is a float value and not decimal, and since the query compares times it's not accurate. It can be fixed by substracting a epsilon to the runtime part, kinda like this: 

SELECT count(runtimepro)+1 FROM `ck_playertimes` WHERE `mapname` = '%s' AND `runtimepro` < %f-(1E-3);

credit to Alerad for this one.