Band-aid
This is more of a band-aid than anything else.

This does allow you to get more than 1 jump in whilst not moving or moving at a very low speed <100u/s however when moving at speed that actually matters then it seems to work as it should.

I spent 10+ mins jumping in a zone and i could not break it.